requirements.txt
# - - - - - - - - - - - - - - - - - - - - - - - - #
# Usage:                                          #
# $ pip3 install -r requirements.txt              #
# or                                              #
# $ python3.6 -m pip install -r requirements.txt  #
# - - - - - - - - - - - - - - - - - - - - - - - - #

# =============================================== #
# MAKE SURE TO ADD THE VERSION SO THAT IF A       #
# PACKAGE'S API CHANGES, WE AREN'T STUCK TRYING   #
# TO FIGURE OUT WHAT WENT WRONG!                  #
# =============================================== #

# ------------------------- #
# Django                    #
# ------------------------- #

    django==2.1.2

    # API Visualise Page
    django-rest-swagger==2.2.0
    django-silk==3.0.1

    # Security
    django-cors-headers==2.4.0
    django-auth-ldap==1.7.0

# ------------------------- #
# Django REST Framework     #
# ------------------------- #

    djangorestframework==3.8.2

    # Requests
    django-filter==2.0.0
    drf-nested-routers==0.91

    # Security:
    django-rest-auth==0.9.3
    djangorestframework-jwt==1.11.0

# ------------------------- #
# Testing & Code Coverage   #
# ------------------------- #

    coverage==4.5.1
    pytest-cov==2.6.0
    pytest-django==3.4.2

# ------------------------- #
# Util                      #
# ------------------------- #

    jsonschema==2.6.0
    uuid==1.30

    # Math
    numpy==1.15.2
    pandas==0.23.4

    # Limited Python Interpreter
    asteval==0.9.13

    # Excel Import - Pandas has capability too
    xlrd==1.1.0
